Fix npm being unavailable

+++ docker/image/console/root/lib/task/skeleton/apply.sh
@@ -2,10 +2,12 @@
 
 function task_skeleton_apply()
 {
-    if [ "${PHP_VERSION:0:3}" == "5.6" ]; then
-      mv /home/build/skeleton/composer-5.6.json /home/build/skeleton/composer.json
-    else
-      rm -f /home/build/skeleton/composer-5.6.json
+    if [ -f /home/build/skeleton/composer-5.6.json ]; then
+        if [ "${PHP_VERSION:0:3}" == "5.6" ]; then
+            mv /home/build/skeleton/composer-5.6.json /home/build/skeleton/composer.json
+        else
+            rm -f /home/build/skeleton/composer-5.6.json
+        fi
     fi
     run "rsync --exclude='*.twig' --ignore-existing -a /home/build/skeleton/ /app/"
 }
